Rename completed_escrows to job_completion
