+++ packages/examples/cvat/exchange-oracle/src/crons/webhooks/recording_oracle.py
@@ -19,7 +19,7 @@ from src.core.types import (
 from src.crons._cron_job import cron_job
 from src.crons.webhooks._common import handle_webhook, process_outgoing_webhooks
 from src.db.utils import ForUpdateParams
-from src.handlers.completed_escrows import handle_escrow_export
+from src.handlers.job_completion import handle_escrow_export
 from src.models.webhook import Webhook
