+++ packages/examples/cvat/exchange-oracle/src/handlers/job_completion.py
